pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-ecology3"
version = "0.1.0"
description = "Agent Ecology v3 - clean-room rebuild"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.6.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "litellm>=1.74.0",
]

[project.scripts]
agent-ecology3 = "agent_ecology3.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = ["-q"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
